requirements.txt
discord.py
python-dotenv
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
pytz
google-generativeai 